requirements.txt
odoorpc==0.6.1
